Move definition of internal types to a single file
